chore: move types to index file
