--- src/background/services/history/types.ts
@@ -1,12 +1,0 @@
-import type { IdentityMetadata, OperationType } from "@src/types";
-
-export interface OperationOptions {
-  identity: {
-    commitment: string;
-    metadata: IdentityMetadata;
-  };
-}
-
-export interface OperationFilter {
-  type: OperationType;
-}
